# fix(engine): isolate netnses so insecure-root-added iptables rules dont leak through the bk cni pool [DEV-4584]