attempt at fixing #8585. 

This PR sets a host name for every netns to avoid buildkits netns pooling, which leaks changes to netnses across the pool. This is only relevant to insecure privileged containers, but can lead to very unpredictable network behavior when users modify the network namespace. 

This fix does come with a perf hit, at least on macos. See below for benchmarks. The perf hit has been mitigated by only avoiding pooling for privileged containers.