@sipsma 2024-10-16
> I suspect this means every WithExec will get a completely fresh netns, when really we want them inheriting their parent, but making them unable to modify it (cloning it, basically?)

I get the desire for that inheritance behavior but for now I'd actually say we should just ensure each exec gets a fresh netns; if a user needs certain settings to be applied it's up to them to do that in each exec.
* Reason being that getting COW-like behavior for netns settings is most likely extremely difficult (if possible at all)

> I also tried setting the CNI pool size to 0. That appears to have no effect- the test fails the same way as it does on main.

Ah, looking a little deeper it appears that when it's 0 they will still be reused but released after some timeout:
https://github.com/moby/buildkit/blob/4b4a9aab1e21a8e3754e30e0b16f3bd35bf52f8f/util/network/cniprovider/cni.go#L217

So I think the quickest way to try out disabling the pool would probably be to set the hostname to a random value if not already set.

Then once we know whether there's any noticeable difference we can go from there.