@cwlbraa 2024-10-16
@sipsma sgtm, I'll rejigger the test and try to poke at performance. any tips on how or what you'd benchmark?