@cwlbraa 2024-10-17
for posterity, here are some benchmark results using TestModule

Main, pooled netns:
`dagger call test specific --run="TestModule" --parallel=12  262.70s user 116.87s system 26% cpu 23:57.99 total`

https://dagger.cloud/dagger/traces/f4a24989481466a34e8301f25b07f8db

With isolated netns, timeout
`dagger call test specific --run="TestModule" --parallel=12  298.42s user 149.93s system 23% cpu 31:52.20 total`

https://dagger.cloud/dagger/traces/769b8a04c4babaa55716a60c7e8a2b2d

With isolated netns, no timeout:
`dagger call test specific --run="TestModule" --parallel=12  291.09s user 139.21s system 26% cpu 26:43.30 total`

https://dagger.cloud/dagger/traces/54370f5c8ac54498e961af0b64c20437