@cwlbraa 2024-10-17
Benchmarking without all of our test weirdness, just spawning many alpine processes:

```
type HelloDagger struct{}

// Returns a container that echoes whatever string argument is provided
func (m *HelloDagger) ContainerEcho(stringArg string) *dagger.Container {
	return dag.Container().From("alpine:latest").WithExec([]string{"echo", stringArg})
}

func (m *HelloDagger) Benchmark(ctx context.Context) error {
	g, ctx := errgroup.WithContext(ctx)

	for i := 0; i < 256; i++ {
		i := i // https://golang.org/doc/faq#closures_and_goroutines
		g.Go(func() error {
			guid := uuid.New().String()
			echoMessage := fmt.Sprintf("Echo call #%d - GUID: %s", i+1, guid)
			_, err := m.ContainerEcho(echoMessage).Stdout(ctx)
			return err
		})
	}

	return g.Wait()
}
```

Using released version:

```hyperfine --warmup 3 "dagger call benchmark"
Benchmark 1: dagger call benchmark
  Time (mean ± σ):     10.767 s ±  1.374 s    [User: 0.342 s, System: 0.127 s]
  Range (min … max):    9.055 s … 13.179 s    10 runs```


With main:
```hyperfine --warmup 3 "~/src/dagger/hack/with-dev dagger call benchmark"
Benchmark 1: ~/src/dagger/hack/with-dev dagger call benchmark
  Time (mean ± σ):     11.581 s ±  2.329 s    [User: 0.344 s, System: 0.129 s]
  Range (min … max):    9.071 s … 15.622 s    10 runs```

With netns isolation:
```hyperfine --warmup 3 "~/src/dagger/hack/with-dev dagger call benchmark"
Benchmark 1: ~/src/dagger/hack/with-dev dagger call benchmark
  Time (mean ± σ):     15.038 s ±  3.349 s    [User: 0.400 s, System: 0.147 s]
  Range (min … max):   10.796 s … 20.001 s    10 runs```


There’s clearly a significant perf penalty with this approach. Running more authoritative benchmarks right now.