@cwlbraa 2024-10-17
with 1 warmup run and 64 runs to clear out noise:

with main:

```sh
Benchmark 1: with-dev dagger call benchmark
  Time (mean ± σ):     11.540 s ±  1.743 s    [User: 0.368 s, System: 0.138 s]
  Range (min … max):    8.807 s … 15.886 s    64 runs
```

with netns isolation:
```sh
Benchmark 1: with-dev dagger call benchmark
  Time (mean ± σ):     13.848 s ±  2.160 s    [User: 0.431 s, System: 0.157 s]
  Range (min … max):   10.767 s … 18.507 s    64 runs
```