@sipsma 2024-10-18
@cwlbraa awesome, thanks for gathering all that, super useful!

In absolute terms, that's not *terrible* but yeah feels like it's worth avoiding.

I'd be okay with only throwing the netns away for privileged execs, which AFAICT would be the only ones capable of making these sort of persistent networking changes anyways. Then the only impacted use cases would be users spinning up *tons* of privileged execs very rapidly. Most known use cases for privileged execs are relatively long-running services, so the tradeoff feels worth it here.

Rough idea to implement:
1. You can check whether it's a privileged exec in `executor_spec.go` w/ `if state.procInfo.Meta.SecurityMode == pb.SecurityMode_INSECURE`
2. For now, probably just ensure the hostname is not `""` if it's privileged (set it to a random value if not already set). I think this should work totally fine
  * The alternative is to upstream something that lets you force the netns to not return to the pool, but that would mostly just look slightly nicer while having the same effect