@cwlbraa 2024-11-05
easy enough, I can't believe I hadn't already considered that mitigation.