@cwlbraa 2024-11-05
due diligence reruns, 10 runs (so we see the noise here, but all else should be equal)

on main
```
Benchmark 1: with-dev dagger call benchmark
  Time (mean ± σ):     14.387 s ±  2.685 s    [User: 0.777 s, System: 0.231 s]
  Range (min … max):   11.071 s … 18.956 s    10 runs
```

with isolation:
```
Benchmark 1: with-dev dagger call benchmark
  Time (mean ± σ):     12.395 s ±  2.294 s    [User: 0.730 s, System: 0.212 s]
  Range (min … max):   10.100 s … 16.482 s    10 runs
```