@cwlbraa 2024-11-08
this appears to [break](https://dagger.cloud/dagger/traces/554123412bcb0bd583a122fc03aa4cec) "TestWithHostnameModuleScoping/works within module" ... I can't tell if that's by design or what, gonna have to do some research on the intent of that test.